Do clean only every 10th build

+++ ci/cibuild.bash
@@ -28,9 +28,14 @@ va_hakija_default_source_path="va-hakija/target/uberjar/hakija-*-standalone.jar"
 va_virkailija_default_source_path="va-virkailija/target/uberjar/virkailija-*-standalone.jar"
 
 function clean() {
-  echo "Running lein clean and emptying all subdirectories with name 'node_modules'"
-  $LEIN modules clean
-  find $PROJECTROOTPATH -depth -type d -name 'node_modules' -exec rm -rf {} \;
+  if [[ "$BUILD_NUMBER" == *0 ]] ;
+  then
+    echo "Running lein clean and emptying all subdirectories with name 'node_modules', (BUILD_NUMBER=$BUILD_NUMBER)"
+    $LEIN modules clean
+    find $PROJECTROOTPATH -depth -type d -name 'node_modules' -exec rm -rf {} \;
+  else
+    echo "Skipping lein clean and emptying all subdirectories with name 'node_modules', (BUILD_NUMBER=$BUILD_NUMBER)"
+  fi
 }
 
 function install_module() {
